minors

+++ src/utils.py
@@ -87,7 +87,7 @@ def plot(image, template, label=None, logit=None, target='save', dest='.'):
 
 @tf.function
 def get_zero_base_label(labels):
-    return tf.divide(tf.add(labels, 1), 2, dtype=tf.int32)
+    return tf.divide(tf.add(labels, 1.0), 2.0)
 
 
 def plot_dataset(dataset, samples, target='save', dest='.'):
